+++ src/CG.Business/Repositories/ApplicationBuilderExtensions.cs
@@ -151,11 +151,15 @@ namespace Microsoft.AspNetCore.Builder
             // Format the name of a target extension method.
             var methodName = $"Use{strategyName}Repositories";
 
+            // Our convention is that the specified extension method can appear in
+            //   two different flavors, one with a configuration parameter, and the
+            //   other without. We'll look for both now.
+
             // Look for specified extension method(s).
             var methods = AppDomain.CurrentDomain.ExtensionMethods(
                 typeof(IApplicationBuilder),
                 methodName,
-                new Type[] { typeof(string) },
+                new Type[] { typeof(IHostEnvironment), typeof(IConfiguration) },
                 assemblyWhiteList,
                 assemblyBlackList
                 );
@@ -174,18 +178,45 @@ namespace Microsoft.AspNetCore.Builder
             }
             else
             {
-                // If we get here we found the assembly but couldn't find a matching
-                //   extension method!
+                // If we get here then we failed to find a 3 parameter version of the
+                //   specified extension method. So we'll now look for a two parameter
+                //   version, in case that exists.
 
-                // Panic!
-                throw new BusinessException(
-                    message: string.Format(
-                        Resources.MethodNotFound,
-                        nameof(UseRepositories),
-                        methodName,
-                        $"{nameof(IApplicationBuilder)},{nameof(IHostEnvironment)},{nameof(IConfiguration)}"
-                        )
+                // Look for specified extension method(s).
+                methods = AppDomain.CurrentDomain.ExtensionMethods(
+                    typeof(IApplicationBuilder),
+                    methodName,
+                    new Type[] { typeof(IHostEnvironment) },
+                    assemblyWhiteList,
+                    assemblyBlackList
                     );
+
+                // Did we find anything?
+                if (methods.Any())
+                {
+                    // We'll use the first matching method.
+                    var method = methods.First();
+
+                    // Invoke the extension method.
+                    method.Invoke(
+                        null,
+                        new object[] { applicationBuilder, hostEnvironment }
+                        );
+                }
+                else
+                {
+                    // If we get here we couldn't find the specified extension method
+                    //   in any combination of parameter(s).
+
+                    // Panic!
+                    throw new BusinessException(
+                        message: string.Format(
+                            Resources.MethodNotFound,
+                            nameof(UseRepositories),
+                            methodName
+                            )
+                        );
+                }
             }
 
             // Return the application builder.